+++ CHANGELOG.md
@@ -1,3 +1,15 @@
+## v1.11.1 (2024-11-29)
+
+### 🐛🚑️ Fixes
+
+- **humbl_compass**: add empty extra field if it doenst exist
+
+## v1.11.0 (2024-11-29)
+
+### ✨ Features
+
+- **humbl_compass**: added humbl_regime recommendations
+
 ## v1.10.1 (2024-11-28)
 
 ### 🐛🚑️ Fixes